+++ src/helpers/find-config.ts
@@ -1,10 +1,6 @@
-import {ConfigNotFoundError, ConfigEmptyError, composeMessage} from '../errors';
-import {GraphQLConfigResult} from '../types';
-import {
-  createCosmiConfig,
-  createCosmiConfigSync,
-  ConfigSearchResult,
-} from './cosmiconfig';
+import { ConfigNotFoundError, ConfigEmptyError, composeMessage } from '../errors';
+import { GraphQLConfigResult } from '../types';
+import { createCosmiConfig, createCosmiConfigSync, ConfigSearchResult } from './cosmiconfig';
 
 const cwd = typeof process !== 'undefined' ? process.cwd() : undefined;
 
@@ -17,11 +13,11 @@ export async function findConfig({
   configName: string;
   legacy?: boolean;
 }): Promise<GraphQLConfigResult> {
-  validate({rootDir});
+  validate({ rootDir });
 
   return resolve({
     rootDir,
-    result: await createCosmiConfig(configName, {legacy}).search(rootDir),
+    result: await createCosmiConfig(configName, { legacy }).search(rootDir),
   });
 }
 
@@ -34,29 +30,23 @@ export function findConfigSync({
   configName: string;
   legacy?: boolean;
 }): GraphQLConfigResult {
-  validate({rootDir});
+  validate({ rootDir });
 
   return resolve({
     rootDir,
-    result: createCosmiConfigSync(configName, {legacy}).search(rootDir),
+    result: createCosmiConfigSync(configName, { legacy }).search(rootDir),
   });
 }
 
 //
 
-function validate({rootDir}: {rootDir: string}) {
+function validate({ rootDir }: { rootDir: string }) {
   if (!rootDir) {
     throw new Error(`Defining a root directory is required`);
   }
 }
 
-function resolve({
-  result,
-  rootDir,
-}: {
-  result?: ConfigSearchResult;
-  rootDir: string;
-}) {
+function resolve({ result, rootDir }: { result?: ConfigSearchResult; rootDir: string }) {
   if (!result) {
     throw new ConfigNotFoundError(
       composeMessage(
@@ -67,12 +57,7 @@ function resolve({
   }
 
   if (result.isEmpty) {
-    throw new ConfigEmptyError(
-      composeMessage(
-        `GraphQL Config file is empty.`,
-        `Please check ${result.filepath}`,
-      ),
-    );
+    throw new ConfigEmptyError(composeMessage(`GraphQL Config file is empty.`, `Please check ${result.filepath}`));
   }
 
   return {